pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxbridge"
version = "0.1.0"
description = "Universal telephony adapter SDK for voice bots - connect any WebSocket voice bot to any telephony platform"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "VoxBridge Contributors" },
]
keywords = ["telephony", "voip", "sip", "websocket", "voice-bot", "twilio", "genesys", "avaya"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Telephony",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO",
]
dependencies = [
    "pydantic>=2.0",
    "websockets>=12.0",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
opus = ["opuslib>=3.0"]
sip = ["pjsua2>=2.13"]
server = ["fastapi>=0.110", "uvicorn>=0.27"]
all = ["voxbridge[opus,server]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[project.scripts]
voxbridge = "voxbridge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["voxbridge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
